feat: 公共方法新增deepEqual判断两个对象是否内容相同
