+++ components/Table/Td.js
@@ -1,5 +1,5 @@
 import Component from '../Component/index'
-import { getStyle, isFunction, isNumeric, isPlainObject, isString } from '../util/index'
+import { deepEqual, getStyle, isFunction, isNumeric, isPlainObject, isString } from '../util/index'
 
 class Td extends Component {
   constructor(props, ...mixins) {
@@ -758,61 +758,6 @@ class Td extends Component {
     this._updateTdData()
   }
 
-  _deepEqual(obj1, obj2) {
-    if (obj1 === obj2) {
-      return true
-    }
-
-    if (typeof obj1 !== typeof obj2) {
-      return false
-    }
-
-    if (Array.isArray(obj1) && Array.isArray(obj2)) {
-      // 如果两个都是数组，比较数组元素
-      return this._compareArrays(obj1, obj2)
-    }
-    if (typeof obj1 === 'object' && obj1 !== null && typeof obj2 === 'object' && obj2 !== null) {
-      // 如果两个都是对象，比较对象的键值对
-      const keys1 = Object.keys(obj1),
-        keys2 = Object.keys(obj2)
-      if (keys1.length !== keys2.length) {
-        return false
-      }
-      for (const key of keys1) {
-        if (
-          !Object.prototype.hasOwnProperty.call(obj2, key) ||
-          !this._deepEqual(obj1[key], obj2[key])
-        ) {
-          return false
-        }
-      }
-      return true
-    }
-    // 基本类型比较
-    return false
-  }
-
-  _compareArrays(arr1, arr2) {
-    // 检查两个数组长度是否相同
-    if (arr1.length !== arr2.length) {
-      return false
-    }
-    // 创建一个包含所有元素的集合
-    const set1 = new Set(arr1.map((item) => JSON.stringify(item)))
-    const set2 = new Set(arr2.map((item) => JSON.stringify(item)))
-    // 比较集合的大小
-    if (set1.size !== set2.size) {
-      return false
-    }
-    // 比较集合中的元素
-    for (const item of set1) {
-      if (!set2.has(item)) {
-        return false
-      }
-    }
-    return true
-  }
-
   _updateTdData() {
     if (!this.editor.validate()) {
       this.table.grid.props.excelMode.onValidateFailed &&
@@ -825,7 +770,7 @@ class Td extends Component {
 
     const newData = this.editor.getValue()
 
-    if (!this._deepEqual(this.props.data, newData)) {
+    if (!deepEqual(this.props.data, newData)) {
       this._onCellValueChange({ newValue: newData, oldValue: this.props.data })
       this.props.data = newData
